# rigweda/backend/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rigweda API
  version: 0.8.0
  description: Independently implemented HR and payroll platform API.
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: Sign in and establish a rotating session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200": { description: Authentication succeeded; refresh token set as an HTTP-only cookie }
        "401": { description: Invalid credentials }
        "423": { description: Account temporarily locked }
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate the refresh session and issue an access token
      responses:
        "200": { description: Session rotated }
        "401": { description: Session missing or invalid }
...
  /auth/me:
    get:
      tags: [Authentication]
      summary: Retrieve the authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current user }
        "401": { description: Authentication required }
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change the current password and revoke other sessions
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 12 }
      responses:
        "204": { description: Password changed }
        "401": { description: Current password is incorrect }
        "422": { description: Password policy validation failed }
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Prepare password-reset instructions without disclosing account existence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: Generic reset response }
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Consume a one-time reset token and revoke all sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, format: password, minLength: 12 }
      responses:
        "204": { description: Password reset }
        "422": { description: Reset token invalid, expired, or already used }
  /auth/sessions:
    get:
      tags: [Authentication]
      summary: List active sessions
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Active sessions }
    delete:
      tags: [Authentication]
      summary: Revoke every session except the current session
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Other sessions revoked }
  /auth/sessions/{sessionId}:
    delete:
      tags: [Authentication]
      summary: Revoke one active session
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Session revoked }
        "404": { description: Session not found }
  /auth/preferences:
    get:
      tags: [Preferences]
      summary: Retrieve personal preferences
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Personal preferences }
    patch:
      tags: [Preferences]
      summary: Update personal preferences
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Updated personal preferences }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...